@@ -18,7 +18,7 @@
     "config": {
         "sort-packages": true,
         "platform": {
-            "php": "8.0.99"
+            "php": "8.1.99"
         },
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true,
@@ -32,28 +32,27 @@
         }
     },
     "require": {
-        "php": "~8.0.0 || ~8.1.0 || ~8.2.0",
+        "php": "~8.1.0 || ~8.2.0",
         "ext-intl": "*",
-        "laminas/laminas-servicemanager": "^3.14.0",
-        "laminas/laminas-stdlib": "^2.7 || ^3.0"
+        "laminas/laminas-servicemanager": "^3.21.0",
+        "laminas/laminas-stdlib": "^3.0"
     },
     "require-dev": {
-        "laminas/laminas-cache": "^3.8",
+        "laminas/laminas-cache": "^3.10.1",
         "laminas/laminas-cache-storage-adapter-memory": "^2.2.0",
-        "laminas/laminas-cache-storage-deprecated-factory": "^1.0.1",
+        "laminas/laminas-cache-storage-deprecated-factory": "^1.1",
         "laminas/laminas-coding-standard": "~2.5.0",
         "laminas/laminas-config": "^3.8.0",
-        "laminas/laminas-eventmanager": "^3.7",
-        "laminas/laminas-filter": "^2.28.1",
-        "laminas/laminas-validator": "^2.28",
-        "laminas/laminas-view": "^2.25",
-        "phpunit/phpunit": "^9.5.26",
-        "psalm/plugin-phpunit": "^0.18.3",
-        "vimeo/psalm": "^5.0.0"
+        "laminas/laminas-eventmanager": "^3.10",
+        "laminas/laminas-filter": "^2.31",
+        "laminas/laminas-validator": "^2.30.1",
+        "laminas/laminas-view": "^2.27",
+        "phpunit/phpunit": "^10.1.3",
+        "psalm/plugin-phpunit": "^0.18.4",
+        "vimeo/psalm": "^5.11"
     },
     "conflict": {
         "laminas/laminas-view": "<2.20.0",
-        "phpspec/prophecy": "<1.9.0",
         "zendframework/zend-i18n": "*"
     },
     "suggest": {
