@@ -7,6 +7,7 @@ namespace LaminasTest\I18n\Filter;
 use Laminas\I18n\Filter\Alpha as AlphaFilter;
 use LaminasTest\I18n\TestCase;
 use Locale;
+use PHPUnit\Framework\Attributes\DataProvider;
 use stdClass;
 
 use function array_keys;
@@ -152,7 +153,7 @@ class AlphaTest extends TestCase
     }
 
     /** @return array<array-key, array{0: mixed}> */
-    public function returnUnfilteredDataProvider(): array
+    public static function returnUnfilteredDataProvider(): array
     {
         return [
             [null],
@@ -161,9 +162,9 @@ class AlphaTest extends TestCase
     }
 
     /**
-     * @dataProvider returnUnfilteredDataProvider
      * @param mixed $input
      */
+    #[DataProvider('returnUnfilteredDataProvider')]
     public function testReturnUnfiltered($input): void
     {
         $filter = new AlphaFilter();
